requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
prometheus-client==0.19.0
python-multipart==0.0.6
email-validator==2.0.0
psutil==5.9.5
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-exporter-otlp==1.37.0
deprecated==1.2.18
# ELK Stack dependencies
elasticsearch==8.11.0
python-json-logger==2.0.7
structlog==23.2.0